+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project are documented in this file.
 
-## [0.3.0] - Unreleased
+## [0.3.0] - 2026-09-12
 
 ### Removed